requirements.txt
streamlit
websocket-client
smartapi-python
pyinstaller==5.11.0
